example/requirements.txt
langchain-core>=0.3.59
langchain-astradb>=0.6.0
langchain-openai>=0.3.16
langchain-graph-retriever>=0.8.0
spacy>=3.8.5
graph-rag-example-helpers>=0.8.0
python-dotenv>=1.1.0
astrapy>=2.0.1
networkx>=3.4.2
immutabledict>=4.2.1
simsimd>=6.2.1 